- Add domain, data, presentation layers - Refactor domain layer
