+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/data/database/AppDatabase.kt
@@ -1,10 +1,10 @@
-package ua.com.koval.andrey.cryptoapp.database
+package ua.com.koval.andrey.cryptoapp.data.database
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import ua.com.koval.andrey.cryptoapp.pojo.CoinPriceInfo
+import ua.com.koval.andrey.cryptoapp.data.model.CoinPriceInfo
 
 @Database(entities = [CoinPriceInfo::class], version = 1, exportSchema = false)
 abstract class AppDatabase: RoomDatabase() {
@@ -13,7 +13,7 @@ abstract class AppDatabase: RoomDatabase() {
         private const val DB_NAME = "main.db"
         private val LOCK = Any()
 
-        fun getInstance(context: Context): AppDatabase{
+        fun getInstance(context: Context): AppDatabase {
             synchronized(LOCK) {
                 db?.let { return it }
                 val instance =
