+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/data/model/CoinPriceInfo.kt
@@ -1,10 +1,10 @@
-package ua.com.koval.andrey.cryptoapp.pojo
+package ua.com.koval.andrey.cryptoapp.data.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
-import ua.com.koval.andrey.cryptoapp.api.ApiFactory.BASE_IMG_URL
+import ua.com.koval.andrey.cryptoapp.data.network.ApiFactory.BASE_IMG_URL
 import ua.com.koval.andrey.cryptoapp.utils.convertTimestampToTime
 
 @Entity(tableName = "full_price_info")
