+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/data/network/ApiFactory.kt
@@ -1,4 +1,4 @@
-package ua.com.koval.andrey.cryptoapp.api
+package ua.com.koval.andrey.cryptoapp.data.network
 
 import hu.akarnokd.rxjava3.retrofit.RxJava3CallAdapterFactory
 import retrofit2.Retrofit
