+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/domain/CoinPriceInfo.kt
@@ -1,0 +1,13 @@
+package ua.com.koval.andrey.cryptoapp.domain
+
+
+data class CoinPriceInfo(
+    val fromSymbol: String,
+    val toSymbol: String?,
+    val price: String?,
+    val lastUpdate: Long?,
+    val highDay: String?,
+    val lowDay: String?,
+    val lastMarket: String?,
+    val imageUrl: String?
+)
