+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/domain/CoinRepository.kt
@@ -1,0 +1,10 @@
+package ua.com.koval.andrey.cryptoapp.domain
+
+import androidx.lifecycle.LiveData
+import ua.com.koval.andrey.cryptoapp.data.model.CoinInfo
+
+interface CoinRepository {
+    fun getCoinInfoList(): LiveData<List<CoinInfo>>
+
+    fun getCoinInfo(fromSymbol: String): LiveData<CoinInfo>
+}
