+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/domain/GetCoinInfoListUseCase.kt
@@ -1,0 +1,7 @@
+package ua.com.koval.andrey.cryptoapp.domain
+
+class GetCoinInfoListUseCase(
+    private val repository: CoinRepository
+    ){
+    operator fun invoke() = repository.getCoinInfoList()
+}
