+++ app/src/main/java/ua/com/koval/andrey/cryptoapp/domain/GetCoinInfoUseCase.kt
@@ -1,0 +1,7 @@
+package ua.com.koval.andrey.cryptoapp.domain
+
+class GetCoinInfoUseCase(
+    private val repository: CoinRepository
+    ){
+    operator fun invoke(fromSymbol: String) = repository.getCoinInfo(fromSymbol)
+}
