+++ app/src/main/res/layout/activity_coin_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.CoinDetailActivity">
+    tools:context=".presentation.CoinDetailActivity">
 
     <ImageView
         android:id="@+id/ivLogoCoin"
