+++ app/src/main/res/layout/activity_coin_price_list.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.CoinPriceListActivity">
+    tools:context=".presentation.CoinPriceListActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvCoinPriceList"
